Added ML Model script and eased the process by adding side conditions

+++ Data-Mining-Scripts/git_checkout_parser.py
@@ -18,30 +18,9 @@ except ImportError:
     os.system("pip3 install xlrd")
     import git 
     import numpy as np
+    from sklearn.preprocessing import LabelEncoder
     import xlrd
     import pandas as pd
-
-# Custom Class for adjusting to new and unseen labels
-class LabelEncoderExt(object):
-    def __init__(self):
-        self.label_encoder = LabelEncoder()
-
-    def fit(self, data_list):
-        self.label_encoder = self.label_encoder.fit(list(data_list) + ['Unknown'])
-        self.classes_ = self.label_encoder.classes_
-
-        return self
-
-    def transform(self, data_list):
-        new_data_list = list(data_list)
-        for unique_item in np.unique(data_list):
-            if unique_item not in self.label_encoder.classes_:
-                new_data_list = ['Unknown' if x==unique_item else x for x in new_data_list]
-
-        return self.label_encoder.transform(new_data_list)
-    
-    def inverse_transform(self, data_list):
-        return self.label_encoder.inverse_transform(data_list)
 
 def cartesian_product(left, right):
 	if right.empty == True:
@@ -49,6 +28,20 @@ def cartesian_product(left, right):
 		right = right.append(pd.Series(0, index = right.columns), ignore_index=True)
 	return pd.concat([pd.concat([left]*len(right)).sort_index().reset_index(drop=True),
        pd.concat([right]*len(left)).reset_index(drop=True) ], 1)
+
+def clean_data(data):
+	req_cols_data = [col for col in data.columns if col.lower()[:7] != 'unnamed']
+	data = data[req_cols_data]
+	data.set_index('index', inplace = True)
+
+	data = pd.concat([data, pd.get_dummies(data[['Author']])], axis=1)
+	data = data.fillna(0)
+	data['Bug_present'] = data['Bug_present'].astype(int)
+	data['is_changed'] = data['is_changed'].astype(int)
+	data['Bug_present'] = data['Bug_present'] & data['is_changed']
+	data.drop(['Commit_Hash', 'Date_Time', 'Author', 'Commit_Message', 'is_changed'], axis = 1, inplace = True)
+	return data
+
 
 parser  =   optparse.OptionParser()
 
@@ -72,8 +65,9 @@ final_data = pd.DataFrame(columns=['index', 'Commit_Hash', 'Author', 'Date_Time'
 training_dir = parent_directory + "/logs/training_data"
 testing_dir = parent_directory + "/logs/testing_data"
 
-xlsx_file_name = path.split('/')[-1] + '_training_data.xlsx'
-final_data.to_excel(training_dir + '/' + xlsx_file_name)
+training_data_file_name = path.split('/')[-1] + '_training_data.xlsx'
+testing_data_file_name = path.split('/')[-1] + '_testing_data.xlsx'
+final_data.to_excel(training_dir + '/' + training_data_file_name)
 print(training_dir)
 
 for i in range(len(df['Commit_Hash'])):
@@ -84,26 +78,31 @@ for i in range(len(df['Commit_Hash'])):
 	# os.system('git checkout ' + commit)
 	subprocess.check_output('git checkout ' + commit, shell = True)
 	print('\n============= Covering Commit #', (i+1), '| Hash:', commit, '=============')
+	new_file_name_metrics = 'metrics' + commit + '.csv'
+	metrics_df_right = []
+	if os.path.isfile(new_file_name_metrics) == True:
+		print("Already found the file, Reading the metrics......")
+		metrics_df_right = pd.read_csv(new_file_name_metrics)
+	else:
+		# Run Sub-script and generate metrics
+		os.chdir(parent_directory)	
+		# os.system("python3 understand_metrics_parser.py " + repo_location)
+		print('Starting directory cleaning and metrics generation.....')
+		subprocess.check_output("python3 understand_metrics_parser.py " + repo_location, shell=True)
+
+		# Changing directory updates file structure, otherwise it'll give FileNotFoundError
+		os.chdir(parent_directory)	
+		os.chdir(repo_location)
+
+		#Create dataframe and continue
+		os.rename('metrics.csv', new_file_name_metrics)
+		metrics_df_right = pd.read_csv(new_file_name_metrics)
+		metrics_df_right = metrics_df_right.loc[metrics_df_right['Kind'] == 'File']
+		metrics_df_right = metrics_df_right[['Name', 'AvgCyclomatic', 'AvgCyclomaticModified', 'AvgCyclomaticStrict', 'AvgEssential', 'AvgLine', 'AvgLineBlank', 'AvgLineCode', 'AvgLineComment', 'CountDeclClass', 'CountDeclClassMethod', 'CountDeclClassVariable', 'CountDeclExecutableUnit', 'CountDeclFunction', 'CountDeclInstanceMethod', 'CountDeclInstanceVariable', 'CountDeclMethod', 'CountDeclMethodDefault', 'CountDeclMethodPrivate', 'CountDeclMethodProtected', 'CountDeclMethodPublic', 'CountLine', 'CountLineBlank', 'CountLineCode', 'CountLineCodeDecl', 'CountLineCodeExe', 'CountLineComment', 'CountSemicolon', 'CountStmt', 'CountStmtDecl', 'CountStmtExe', 'MaxCyclomatic', 'MaxCyclomaticModified', 'MaxCyclomaticStrict', 'MaxEssential', 'MaxNesting', 'RatioCommentToCode', 'SumCyclomatic', 'SumCyclomaticModified', 'SumCyclomaticStrict', 'SumEssential']]
+		metrics_df_right.to_csv(new_file_name_metrics)
+
 	all_commits_left = df.loc[df['Commit_Hash'] == commit]
 	all_commits_left['index'] = (i+1)
-
-	# Run Sub-script and generate metrics
-	os.chdir(parent_directory)	
-	# os.system("python3 understand_metrics_parser.py " + repo_location)
-	print('Starting directory cleaning and metrics generation.....')
-	subprocess.check_output("python3 understand_metrics_parser.py " + repo_location, shell=True)
-
-	# Changing directory updates file structure, otherwise it'll give FileNotFoundError
-	os.chdir(parent_directory)	
-	os.chdir(repo_location)
-
-	#Create dataframe and continue
-	new_file_name_metrics = 'metrics' + commit + '.csv'
-	os.rename('metrics.csv', new_file_name_metrics)
-	metrics_df_right = pd.read_csv(new_file_name_metrics)
-	metrics_df_right = metrics_df_right.loc[metrics_df_right['Kind'] == 'File']
-	metrics_df_right = metrics_df_right[['Name', 'AvgCyclomatic', 'AvgCyclomaticModified', 'AvgCyclomaticStrict', 'AvgEssential', 'AvgLine', 'AvgLineBlank', 'AvgLineCode', 'AvgLineComment', 'CountDeclClass', 'CountDeclClassMethod', 'CountDeclClassVariable', 'CountDeclExecutableUnit', 'CountDeclFunction', 'CountDeclInstanceMethod', 'CountDeclInstanceVariable', 'CountDeclMethod', 'CountDeclMethodDefault', 'CountDeclMethodPrivate', 'CountDeclMethodProtected', 'CountDeclMethodPublic', 'CountLine', 'CountLineBlank', 'CountLineCode', 'CountLineCodeDecl', 'CountLineCodeExe', 'CountLineComment', 'CountSemicolon', 'CountStmt', 'CountStmtDecl', 'CountStmtExe', 'MaxCyclomatic', 'MaxCyclomaticModified', 'MaxCyclomaticStrict', 'MaxEssential', 'MaxNesting', 'RatioCommentToCode', 'SumCyclomatic', 'SumCyclomaticModified', 'SumCyclomaticStrict', 'SumEssential']]
-	metrics_df_right.to_csv(new_file_name_metrics)
 
 	#Part where we add per file difference column named "is_chagned"
 	if i == 0:
@@ -131,21 +130,24 @@ for i in range(len(df['Commit_Hash'])):
 	temp_df = temp_df[req_cols_temp_df]
 	# temp_df.to_excel(str(i) + '.xlsx') #This is the cartesian product excel if you like to see
 
-	final_data = pd.read_excel(training_dir + '/' + xlsx_file_name)
+	final_data = pd.read_excel(training_dir + '/' + training_data_file_name)
 	req_cols_final_data = [col for col in final_data.columns if col.lower()[:7] != 'unnamed']
 	final_data = final_data[req_cols_final_data]
 	print('Final Data before append: ', final_data.shape)
 	final_data = final_data.append(temp_df)
 	print('Final Data after append: ', final_data.shape)
 	print('Temp DF: ', temp_df.shape)
-	final_data.to_excel(training_dir + '/' + xlsx_file_name)
+	final_data.to_excel(training_dir + '/' + training_data_file_name)
 
 	if i == len(df['Commit_Hash'])-1:
-		temp_df.to_excel(testing_dir + '/' + path.split('/')[-1] + '_testing_data.xlsx')
+		temp_df = clean_data(temp_df)
+		# clean_data(testing_dir + '/' + testing_data_file_name)
+		temp_df.to_excel(testing_dir + '/' + testing_data_file_name)
 
 
 # Drop Commit_Hash and is_changed columns
-data = pd.read_excel(training_dir + '/' + xlsx_file_name)
+# clean_data(training_dir + '/' + training_data_file_name)
+data = pd.read_excel(training_dir + '/' + training_data_file_name)
 req_cols_data = [col for col in data.columns if col.lower()[:7] != 'unnamed']
 data = data[req_cols_data]
 data.set_index('index', inplace = True)
@@ -156,4 +158,4 @@ data['Bug_present'] = data['Bug_present'].astype(int)
 data['is_changed'] = data['is_changed'].astype(int)
 data['Bug_present'] = data['Bug_present'] & data['is_changed']
 data.drop(['Commit_Hash', 'Date_Time', 'Author', 'Commit_Message', 'is_changed'], axis = 1, inplace = True)
-data.to_excel(training_dir + '/' + xlsx_file_name)
+data.to_excel(training_dir + '/' + training_data_file_name)
